% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flattenlist.R
\name{flattenlist}
\alias{flattenlist}
\title{FUNCTION_TITLE}
\usage{
flattenlist(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
