% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_path.R
\name{shiny_path}
\alias{shiny_path}
\title{FUNCTION_TITLE}
\usage{
shiny_path(m)
}
\arguments{
\item{m}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 library(ShinyTester)
 
 x <- ShinyTester::ShinyHierarchy(system.file('shinyexample',package='shinycovr'))
 x$x$edges
 x$x$nodes
 
 # possible reactive paths
 shiny_path(x)
 }
}
}
\seealso{
\code{\link[igraph]{as_data_frame}},\code{\link[igraph]{as_adjacency_matrix}}
 \code{\link[stats]{setNames}}
}
