% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{traverse}
\alias{traverse}
\title{FUNCTION_TITLE}
\usage{
traverse(parent, adj, chain = NULL)
}
\arguments{
\item{parent}{PARAM_DESCRIPTION}

\item{adj}{PARAM_DESCRIPTION}

\item{chain}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{setNames}}
}
